Updating page styling
Updating modal card fonts, footer, project page, and file download. Issue with footer:  if there are 3 or fewer projects, the footer will appear high up on the All Projects page; if there are 4 or more projects, then the footer will be at the bottom as expected. This does not affect any other pages.